Add static_cast's from iterator diff type to int

+++ library/bound.hpp
@@ -11,22 +11,24 @@
 // the sub-range of elements that "equal" to val.
 template<typename Sequence, typename Value>
 int lb(const Sequence &seq, const Value &val) {
-  return std::lower_bound(std::begin(seq), std::end(seq), val)
-      - std::begin(seq);
+  return static_cast<int>(std::lower_bound(std::begin(seq), std::end(seq), val)
+                          - std::begin(seq));
 }
 template<typename Sequence, typename Value>
 int ub(const Sequence &seq, const Value &val) {
-  return std::upper_bound(std::begin(seq), std::end(seq), val)
-      - std::begin(seq);
+  return static_cast<int>(std::upper_bound(std::begin(seq), std::end(seq), val)
+                          - std::begin(seq));
 }
 template<typename Sequence, typename Value, typename Compare>
 int lb(const Sequence &seq, const Value &val, Compare comp) {
-  return std::lower_bound(std::begin(seq), std::end(seq), val, comp)
-      - std::begin(seq);
+  return static_cast<int>(
+      std::lower_bound(std::begin(seq), std::end(seq), val, comp)
+      - std::begin(seq));
 }
 template<typename Sequence, typename Value, typename Compare>
 int ub(const Sequence &seq, const Value &val, Compare comp) {
-  return std::upper_bound(std::begin(seq), std::end(seq), val, comp)
-      - std::begin(seq);
+  return static_cast<int>(
+      std::upper_bound(std::begin(seq), std::end(seq), val, comp)
+      - std::begin(seq));
 }
 #endif// JHELPER_EXAMPLE_PROJECT_LIBRARY_BOUND_HPP_
